* setup_mkey.c (krb5_db_setup_mkey_name): Use size_t instead of         int for lengths.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@15069 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/kdb/setup_mkey.c
@@ -47,8 +47,8 @@ krb5_db_setup_mkey_name(context, keyname, realm, fullname, principal)
     krb5_principal *principal;
 {
     krb5_error_code retval;
-    int keylen;
-    int rlen = strlen(realm);
+    size_t keylen;
+    size_t rlen = strlen(realm);
     char *fname;
     
     if (!keyname)
